@@ -11,7 +11,7 @@
         }
     ],
     "require": {
-        "php": ">=7.3",
+        "php": ">=7.4",
         "guzzlehttp/guzzle": "^6.5.5|^7.0",
         "illuminate/config": "^7.0|^8.0",
         "illuminate/support": "^7.0|^8.0",
